+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## 2.1.0 - Unreleased
+### Added
+* Added `django_flexquery.contrib.for_user.ForUserFlexQuery`
 ### Changed
 * Changed directory structure of unittests